Rename module and template
