+++ frontend/src/modules/TornadoChart/queryHooks.tsx
