+++ frontend/src/modules/TornadoChart/sensitivityResponseCalculator.ts
