+++ frontend/src/modules/TornadoChart/sensitivityTable.tsx
