+++ frontend/src/modules/TornadoChart/settings.tsx
