+++ frontend/src/modules/TornadoChart/view.tsx
